requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.3.1
djangorestframework==3.15.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
setuptools==76.0.0
sqlparse==0.5.3
wheel==0.45.1

dotenv~=0.9.9
python-dotenv~=1.1.0
django-filter~=25.1
sendgrid~=6.11.0
urllib3~=2.4.0